fixing structarr to preserve indicies dtype

+++ vsm/structarr.py
@@ -85,7 +85,7 @@ def enum_array(arr, indices=[], field_name='i'):
     if len(indices) == 0:
         indices = np.arange(arr.size, dtype=np.int)
     else:
-        indices = np.array(indices, dtype=np.int)
+        indices = np.array(indices)
     return zip_arr(indices, arr, field_names=[field_name, 'value'])
 
 
@@ -127,7 +127,7 @@ def enum_matrix(arr, axis=0, indices=[], field_name='i'):
         indices = np.arange(arr.shape[1], dtype=np.int)
     
     if type(indices) == list:
-        indices = np.array(indices, dtype=np.int)
+        indices = np.array(indices)
     ind = np.array([indices.copy() for i in xrange(arr.shape[0])])
     dt = [(field_name, indices.dtype), ('value', arr.dtype)]
     mt = zip_arr(ind, arr, field_names=[field_name, 'value'])
@@ -187,7 +187,7 @@ def enum_sort(arr, indices=[], field_name='i', filter_nan=False):
     if len(indices) == 0:
         indices = np.arange(arr.shape[0], dtype=np.int)
     else:
-        indices = np.array(indices, dtype=np.int)
+        indices = np.array(indices)
 	
     dt = [(field_name, indices.dtype), ('value', arr.dtype)]
